cis129_lab03_coffeeShop.py
This program is a coffee simulator that requires the user to input options based on how a real coffee shop would work with a menu and a cashier.  In return the code will give you a receipt with the price of each item and tax to show the total cost.